fixed the GA with help of 2 unit tests

+++ evolutionaryComputation/practicals/code/src/main/scala/nl/jappieklooster/hw/ec/Member.scala
@@ -5,19 +5,19 @@ import org.slf4j.LoggerFactory
 import scala.util.Random
 
 trait IMember extends IHasFitness with Genetic
-case class Member(valuation:Int, gen:String) extends IMember{
+case class Member(valuation:Float, gen:String) extends IMember{
 	override def fitness = valuation
 	override def genes = gen
 }
 
 object MemberFactories{
-	def tightlyLinked(func:String=>Int)(str:String):IMember = Member(func(str), str)
-	def randomlyLinked(instructions:TraversableOnce[Int])(func:String=>Int)(str:String):IMember = {
+	def tightlyLinked(func:String=>Float)(str:String):IMember = Member(func(str), str)
+	def randomlyLinked(instructions:TraversableOnce[Int])(func:String=>Float)(str:String):IMember = {
 		val genes = instructions.map(inx => str.charAt(inx)).mkString
 		Member(func(str), str)
 	}
 	val log = LoggerFactory.getLogger(this.getClass)
-	def withCoinfliptimesMutation(creationFunc:(String=>Int) => String => IMember, random:Random)(func:String=>Int)(str:String):IMember = {
+	def withCoinfliptimesMutation(creationFunc:(String=>Float) => String => IMember, random:Random)(func:String=>Float)(str:String):IMember = {
 		def countFlips(integer:Integer):Integer = {
 			if(random.nextBoolean()){
 				return countFlips(integer+1)
